+++ src/main/java/woowa/camp/jspcafe/configuration/AppContextListener.java
@@ -6,8 +6,8 @@ import jakarta.servlet.ServletContextListener;
 import jakarta.servlet.annotation.WebListener;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import woowa.camp.jspcafe.repository.ArticleRepository;
-import woowa.camp.jspcafe.repository.InMemoryArticleRepository;
+import woowa.camp.jspcafe.repository.article.ArticleRepository;
+import woowa.camp.jspcafe.repository.article.InMemoryArticleRepository;
 import woowa.camp.jspcafe.repository.InMemoryUserRepository;
 import woowa.camp.jspcafe.repository.UserRepository;
 import woowa.camp.jspcafe.service.ArticleService;
