+++ src/main/java/woowa/camp/jspcafe/repository/article/ArticleRepository.java
@@ -1,4 +1,4 @@
-package woowa.camp.jspcafe.repository;
+package woowa.camp.jspcafe.repository.article;
 
 import java.util.List;
 import java.util.Optional;
