+++ src/test/java/woowa/camp/jspcafe/service/ArticleServiceTest.java
@@ -14,8 +14,8 @@ import woowa.camp.jspcafe.domain.User;
 import woowa.camp.jspcafe.domain.exception.ArticleException;
 import woowa.camp.jspcafe.fixture.ArticleFixture;
 import woowa.camp.jspcafe.fixture.UserFixture;
-import woowa.camp.jspcafe.repository.ArticleRepository;
-import woowa.camp.jspcafe.repository.InMemoryArticleRepository;
+import woowa.camp.jspcafe.repository.article.ArticleRepository;
+import woowa.camp.jspcafe.repository.article.InMemoryArticleRepository;
 import woowa.camp.jspcafe.repository.InMemoryUserRepository;
 import woowa.camp.jspcafe.repository.UserRepository;
 import woowa.camp.jspcafe.service.dto.ArticleDetailsResponse;
